refactor(front>mobile): Moved `PlatformCard` into `widget` folder.
