+++ frontend/mobile/lib/presentation/new_area/choose_trigger_platform_page.dart
@@ -4,6 +4,7 @@ import 'package:area/layout/main_page_layout.dart';
 import 'package:area/model/platform_model.dart';
 import 'package:area/widget/appbar_button.dart';
 import 'package:area/widget/clickable_frame.dart';
+import 'package:area/widget/platform_card.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -59,7 +60,16 @@ class _ChooseTriggerPlatformPageState extends ConsumerState<ChooseTriggerPlatfor
                   final width = (MediaQuery.of(context).size.width / 2) - 20 - 10;
                   return SizedBox(
                     width: width,
-                    child: PlatformCard(platform: e),
+                    child: PlatformCard(
+                      platform: e,
+                      available: e.triggers.length,
+                      onTap: () {
+                        ref.read(areaModalProvider.notifier)
+                          ..setActionPlatform(e)
+                          ..setTrigger(null);
+                        // context.pop();
+                      },
+                    ),
                   );
                 })
                 .toList(),
@@ -85,68 +95,3 @@ class _ChooseTriggerPlatformPageState extends ConsumerState<ChooseTriggerPlatfor
   }
 
 }
-
-class PlatformCard extends ConsumerWidget {
-
-  final PlatformModel platform;
-
-  const PlatformCard({
-    required this.platform,
-    super.key
-  });
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final textTheme = Theme.of(context).textTheme;
-
-    return ClickableFrame(
-      padding: const EdgeInsets.all(20),
-      onTap: () {
-        ref.read(areaModalProvider.notifier)
-          ..setActionPlatform(platform)
-          ..setTrigger(null);
-        // context.pop();
-      },
-      child: Stack(
-        children: [
-          Column(
-            mainAxisSize: MainAxisSize.min,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Container(
-                height: 48,
-                width: 48,
-                decoration: BoxDecoration(
-                  shape: BoxShape.circle,
-                ),
-                child: Placeholder(), // TODO: Put platform icon
-              ),
-              Gap(5),
-              Text(
-                platform.name,
-                style: textTheme.titleMedium?.copyWith(
-                  fontSize: 20
-                ),
-              ),
-              Text(
-                '${platform.actions.length} available',
-                style: textTheme.bodyMedium,
-              )
-            ],
-          ),
-          Positioned(
-            top: 0,
-            right: 0,
-            child: Icon(
-              Icons.person_off_rounded,
-              color: Colors.red,
-              size: 20,
-            )
-          )
-        ],
-      )
-    );
-  }
-
-}
-
